SmartIFSyn-test/requirements.txt
slither-analyzer==0.10.2
solc-select==1.1.0
tqdm==4.66.4
z3-solver==4.13.0.0